interpol-watcher/webserver/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
pika==1.3.2
python-dotenv==1.0.0
eventlet==0.35.1
